Move `RenderingStates` enum to src/annotator/anchoring/pdf
Move `RenderingStates` enum from `pdfjs-rendering-states.js` to
`src/annotator/anchoring/pdf`.

This is part of an exercise to declutter the root of src/annotator.
