--- src/annotator/pdfjs-rendering-states.js
@@ -1,15 +1,0 @@
-/**
- * Enum values for page rendering states (IRenderableView#renderingState)
- * in PDF.js. Taken from web/pdf_rendering_queue.js in the PDF.js library.
- *
- * Reproduced here because this enum is not exported consistently across
- * different versions of PDF.js
- */
-const RenderingStates = {
-  INITIAL: 0,
-  RUNNING: 1,
-  PAUSED: 2,
-  FINISHED: 3,
-};
-
-export default RenderingStates;
